Extract vms to their own files
