+++ WooCommerce/Classes/ViewRelated/Dashboard/Settings/Webhooks/WebhookRowViewModel.swift
@@ -1,0 +1,11 @@
+import Foundation
+import Yosemite
+
+struct WebhookRowViewModel: Identifiable {
+    var id = UUID()
+    let webhook: Webhook
+
+    init(webhook: Webhook) {
+        self.webhook = webhook
+    }
+}
